DRAM controller magic values cleanup
This patch tries to cleanup as much of the magic values in dram.c as
possible. Obviously for some registers we have no information so cannot
fill those in. No actual code changes took place. This patch is however
based on the most recent dram.c; e.g. from A20.

Signed-off-by: Oliver Schinagl <oliver@schinagl.nl>